+++ CHANGELOG.md
@@ -4,10 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+### Added
+- Configurable log and message format
+
 ### Changed
 - Replaced database migration framework - issue #115
 
     (please, follow the [custom upgrade process](http://storekeeper.readthedocs.org/en/v0.3.0/upgrade.html#upgrade-from-v0-2-1-to-v0-3-0))
+
+### Fixed
+- Can not logging errors via email when message contains UTF-8 characters - issue #126
 
 
 ## [0.2.1] - 2015-11-09